nvme: refactor namespaces to support non-gendisk devices
With LightNVM enabled namespaces, the gendisk structure is not exposed
to the user. This limits LightNVM users to access NVMe device driver
specific sysfs entries, and binding LightNVM namespace geometry to the
specific device.

Refactor the revalidation process, so that a namespace, instead of a
gendisk, is revalidated. This later allows patches to wire up the sysfs
entries up to a non-gendisk namespace.

Signed-off-by: Matias Bjørling <m@bjorling.me>