lightnvm: do not load l2p table on host-based mode
If the device reports that it is not operating in hybrid mode, do not
attempt to load the l2p table.

Also, in the case that the l2p table is loaded, check if the loaded
blocks with unknown type are indeed free blocks before decreasing the
free block count.

Signed-off-by: Javier González <javier@cnexlabs.com>