pblk: Fixed typos and rephrased sections.
Additional input:

The section on the write buffer could benefit by a description similar to
circular-buffers.txt, possibly refer to circular-buffers.txt. like

"pblk write buffer is a ring buffer (See Documentation/circular-buffers.txt)"

Then describe how the pblk write buffer is different from the generic buffer.
This allows the reader to apply their prior knowledge on circular buffers and
possibly their experience in using the "built-in" circular buffer.

If the head/tail are just like the usual head/tail, then not much needs to be
said. But, if assumptions/usage/criteria for adding/removing items
differ then it should be emphasized, otherwise the reader will just assume that
nothing special is going on.

The "circular-buffers.txt" has a very clean description of owner-ship of
pointers and responsibility e.g. producers->head, consumer->tail.
The pblk ring write buffer could benefit from a similar clean description.
If possible, then relate to those terms like describing:

Who has ownership of sync/flush/l2p_update? Does the usage of
head/tail/sync/flush/l2p_update fit into the consumer/producer role of the
generic buffer or are the roles mixed?

Who, what and why produces items in the ring buffer?
Who, what and why consumes items in the ring buffer?

Also, the generic buffer allows for simultanous produce/consume, it would be
nice with notes on how multi-producer/consumer is obtained in the pblk write
buffer. Are they serialized? Locks? Lockless algorithms? Assumptions?

The buffer is also used as a cache, this should be emphasized and described
in closer detail.

I don't know but I suspect that elements undergo state changes while in the
buffer. These would benefit from a description. I have not grasped the code yet
but it seems like the ring buffer conceptually works as:

 * A write buffer
 * A work queue
 * A read cache

When reading the docs the initial impression is that is it just a write buffer,
which fills up until size thresholds are met and then emptied.
But it seems like a hole lot more is going on :)